requirements.txt
python-vlc
mutagen